Port hors_concours integration tests to MiniTest
